+++ config.json
@@ -17,9 +17,9 @@
         "scripts.md": { "disabled": false, "lab": "lab9.py", "server": "server9.py", "runtime": "runtime9.js" },
         "security.md": { "disabled": false, "lab": "lab10.py", "server": "server10.py", "runtime": "runtime10.js" },
 
-        "reflow.md": { "disabled": true, "lab": "lab11.py" },
-        "visual-effects.md": { "disabled": true, "lab": "lab12.py" },
-        "scheduling-and-threading.md": { "disabled": true, "lab": "lab13.py" },
+        "visual-effects.md": { "disabled": true, "lab": "lab11.py" },
+        "scheduling-and-threading.md": { "disabled": true, "lab": "lab12.py" },
+        "reflow.md": { "disabled": true, "lab": "lab13.py" },
 
         "skipped.md": { "disabled": true },
         "change.md": { "disabled": true },
