+++ src/lab12.py
@@ -1,1183 +1,1255 @@
 """
 This file compiles the code in Web Browser Engineering,
-up to and including Chapter 12 (Adding Visual Effects),
+up to and including Chapter 11 (Keeping Data Private),
 without exercises.
 """
 
-import ctypes
+import argparse
 import dukpy
-import io
-import math
-from sdl2 import *
-import skia
+import functools
 import socket
 import ssl
-import urllib.parse
-from PIL import Image
-from lab4 import print_tree
-from lab4 import Element
-from lab4 import Text
-from lab4 import HTMLParser
-from lab6 import cascade_priority
-from lab6 import layout_mode
-from lab6 import resolve_url
-from lab6 import tree_to_list
-from lab6 import INHERITED_PROPERTIES
-from lab6 import compute_style
-from lab6 import TagSelector
-from lab6 import DescendantSelector
-from lab10 import url_origin
-from lab10 import JSContext
+import time
+import threading
+import tkinter
+import tkinter.font
+from lab10 import request
 
-def request(url, headers={}, payload=None):
-    scheme, url = url.split("://", 1)
-    assert scheme in ["http", "https"], \
-        "Unknown scheme {}".format(scheme)
+class Timer:
+    def __init__(self):
+        self.phase = None
+        self.time = None
+        self.accumulated = 0
 
-    host, path = url.split("/", 1)
-    path = "/" + path
-    port = 80 if scheme == "http" else 443
+    def reset(self):
+        self.accumulated = 0
 
-    if ":" in host:
-        host, port = host.split(":", 1)
-        port = int(port)
+    def start(self, name):
+        if self.phase: self.stop()
+        self.phase = name
+        self.time = time.time()
 
-    s = socket.socket(
-        family=socket.AF_INET,
-        type=socket.SOCK_STREAM,
-        proto=socket.IPPROTO_TCP,
-    )
-    s.connect((host, port))
-    if scheme == "https":
-        ctx = ssl.create_default_context()
-        s = ctx.wrap_socket(s, server_hostname=host)
+    def stop(self):
+        dt = time.time() - self.time
+        print("[{:>10.6f}] {}".format(dt, self.phase))
+        self.phase = None
+        self.accumulated += dt
 
-    method = "POST" if payload else "GET"
-    body = "{} {} HTTP/1.0\r\n".format(method, path)
-    body += "Host: {}\r\n".format(host)
-    for header, value in headers.items():
-        body += "{}: {}\r\n".format(header, value)
-    if payload:
-        content_length = len(payload.encode("utf8"))
-        body += "Content-Length: {}\r\n".format(content_length)
-    body += "\r\n" + (payload or "")
-    s.send(body.encode("utf8"))
+    def print_accumulated(self):
+        print("[{:>10.6f}] {}\n".format(self.accumulated, "Total"))
 
-    response = s.makefile("b")
+def url_origin(url):
+    return "/".join(url.split("/")[:3])
 
-    statusline = response.readline().decode("utf8")
-    version, status, explanation = statusline.split(" ", 2)
-    assert status == "200", "{}: {}".format(status, explanation)
-
-    headers = {}
-    while True:
-        line = response.readline().decode("utf8")
-        if line == "\r\n": break
-        header, value = line.split(":", 1)
-        headers[header.lower()] = value.strip()
-
-    if headers.get(
-        'content-type',
-        'application/octet-stream').startswith("text"):
-        body = response.read().decode("utf8")
-    else:
-        body = response.read()
-
-    s.close()
-
-    return headers, body
-
-def color_to_sk_color(color):
-    if color == "white":
-        return skia.ColorWHITE
-    elif color == "lightblue":
-        return skia.ColorSetARGB(0xFF, 0xAD, 0xD8, 0xE6)
-    elif color == "orange":
-        return skia.ColorSetARGB(0xFF, 0xFF, 0xA5, 0x00)
-    elif color == "blue":
-        return skia.ColorBLUE
-    else:
-        return skia.ColorBLACK
-
-def parse_rotation_transform(transform_str):
-    left_paren = transform_str.find('(')
-    right_paren = transform_str.find('deg)')
-    return float(transform_str[left_paren + 1:right_paren])
-
-def parse_blend_mode(blend_mode_str):
-    if blend_mode_str == "multiply":
-        return skia.BlendMode.kMultiply
-    elif blend_mode_str == "difference":
-        return skia.BlendMode.kDifference
-    else:
-        return skia.BlendMode.kSrcOver
-
-def parse_clip_path(clip_path_str):
-    if clip_path_str.find("circle") != 0:
-        return None
-    return int(clip_path_str[7:][:-2])
-
-class Rasterizer:
-    def __init__(self, surface):
-        self.surface = surface
-
-    def clear(self, color):
-        with self.surface as canvas:
-            canvas.clear(color)
-
-    def draw_rect(self, rect,
-        fill=None, width=1):
-        paint = skia.Paint()
-        if fill:
-            paint.setStrokeWidth(width);
-            paint.setColor(color_to_sk_color(fill))
-        else:
-            paint.setStyle(skia.Paint.kStroke_Style)
-            paint.setStrokeWidth(1);
-            paint.setColor(skia.ColorBLACK)
-        with self.surface as canvas:
-            canvas.drawRect(rect, paint)
-
-    def draw_polyline(self, x1, y1, x2, y2, x3=None,
-        y3=None, fill=False):
-        path = skia.Path()
-        path.moveTo(x1, y1)
-        path.lineTo(x2, y2)
-        if x3:
-            path.lineTo(x3, y3)
-        paint = skia.Paint()
-        paint.setColor(skia.ColorBLACK)
-        if fill:
-            paint.setStyle(skia.Paint.kFill_Style)
-        else:
-            paint.setStyle(skia.Paint.kStroke_Style)
-        paint.setStrokeWidth(1);
-        with self.surface as canvas:
-            canvas.drawPath(path, paint)
-
-    def draw_text(self, x, y, text, font, color=None):
-        paint = skia.Paint(
-            AntiAlias=True, Color=color_to_sk_color(color))
-        with self.surface as canvas:
-            canvas.drawString(
-                text, x, y - font.getMetrics().fAscent,
-                font, paint)
-
-def linespace(font):
-    metrics = font.getMetrics()
-    return metrics.fDescent - metrics.fAscent
-
-class SaveLayer:
-    def __init__(self, sk_paint, rect):
-        self.sk_paint = sk_paint
-        self.rect = rect
-
-    def execute(self, scroll, rasterizer):
-        with rasterizer.surface as canvas:
-            canvas.saveLayer(paint=self.sk_paint)
-
-class Save:
-    def __init__(self, rect):
-        self.rect = rect
-
-    def execute(self, scroll, rasterizer):
-        with rasterizer.surface as canvas:
-            canvas.save()
-
-class Restore:
-    def __init__(self, rect):
-        self.rect = rect
-
-    def execute(self, scroll, rasterizer):
-        with rasterizer.surface as canvas:
-            canvas.restore()
-
-class CircleMask:
-    def __init__(self, cx, cy, radius, rect):
-        self.cx = cx
-        self.cy = cy
-        self.radius = radius
-        self.rect = rect
-
-    def execute(self, scroll, rasterizer):
-        with rasterizer.surface as canvas:
-            canvas.saveLayer(paint=skia.Paint(
-                Alphaf=1.0, BlendMode=skia.kDstIn))
-            canvas.drawCircle(
-                self.cx, self.cy - scroll,
-                self.radius, skia.Paint(Color=skia.ColorWHITE))
-
-def center_point(rect):
-    return (rect.left() + (rect.right() - rect.left()) / 2,
-        rect.top() + (rect.bottom() - rect.top()) / 2)
-
-class Rotate:
-    def __init__(self, degrees, rect):
-        self.degrees = degrees
-        self.rect = rect
-
-    def execute(self, scroll, rasterizer):
-        paint_rect = skia.Rect.MakeLTRB(
-            self.rect.left(), self.rect.top() - scroll,
-            self.rect.right(), self.rect.bottom() - scroll)
-        (center_x, center_y) = center_point(paint_rect)
-        with rasterizer.surface as canvas:
-            canvas.translate(-center_x, -center_y)
-            canvas.rotate(self.degrees)
-            canvas.translate(center_x, center_y)
-
-class DrawText:
-    def __init__(self, x1, y1, text, font, color):
-        self.font = font
-        self.rect = skia.Rect.MakeLTRB(x1, y1, x1, y1 + linespace(self.font))
+class Text:
+    def __init__(self, text):
         self.text = text
-        self.color = color
-
-    def execute(self, scroll, rasterizer):
-        rasterizer.draw_text(
-            self.rect.left(), self.rect.top() - scroll,
-            self.text,
-            self.font,
-            self.color,
-        )
 
     def __repr__(self):
-        return "DrawText(text={})".format(self.text)
+        return "\"" + self.text.replace("\n", "\\n") + "\""
 
-class DrawRect:
-    def __init__(self, rect, color):
-        self.rect = rect
-        self.color = color
+SELF_CLOSING_TAGS = [
+    "area", "base", "br", "col", "embed", "hr", "img", "input",
+    "link", "meta", "param", "source", "track", "wbr",
+]
 
-    def execute(self, scroll, rasterizer):
-        rasterizer.draw_rect(
-            skia.Rect.MakeLTRB(
-                self.rect.left(), self.rect.top() - scroll,
-                self.rect.right(), self.rect.bottom() - scroll),
-            width=0,
-            fill=self.color,
-        )
+class Tag:
+    def __init__(self, text):
+        parts = text.split()
+        self.tag = parts[0].lower()
+        self.attributes = {}
+        for attrpair in parts[1:]:
+            if "=" in attrpair:
+                key, value = attrpair.split("=", 1)
+                if len(value) > 2 and value[0] in ["'", "\""]:
+                    value = value[1:-1]
+                self.attributes[key.lower()] = value
+            else:
+                self.attributes[attrpair.lower()] = ""
 
     def __repr__(self):
-        (left, top, right, bottom) = tuple(self.rect)
-        return "DrawRect(top={} left={} bottom={} right={} color={})".format(
-            left, top, right, bottom, self.color)
+        return "<" + self.tag + ">"
 
-class ClipRect:
-    def __init__(self, rect):
-        self.rect = rect
+def lex(body):
+    out = []
+    text = ""
+    in_tag = False
+    for c in body:
+        if c == "<":
+            in_tag = True
+            if text: out.append(Text(text))
+            text = ""
+        elif c == ">":
+            in_tag = False
+            out.append(Tag(text))
+            text = ""
+        else:
+            text += c
+    if not in_tag and text:
+        out.append(Text(text))
+    return out
 
-    def execute(self, scroll, rasterizer):
-        with rasterizer.surface as canvas:
-            canvas.clipRect(skia.Rect.MakeLTRB(
-                self.rect.left(), self.rect.top() - scroll,
-                self.rect.right(), self.rect.bottom() - scroll))
+class ElementNode:
+    def __init__(self, tag, attributes):
+        self.tag = tag
+        self.attributes = attributes
+        self.children = [] 
+        self.style = {}
+        for pair in self.attributes.get("style", "").split(";"):
+            if ":" not in pair: continue
+            prop, val = pair.split(":")
+            self.style[prop.strip().lower()] = val.strip()
 
-class ClipRRect:
-    def __init__(self, rect, radius):
-        self.rect = rect
-        self.radius = radius
+    def __repr__(self):
+        return "<" + self.tag + ">"
 
-    def execute(self, scroll, rasterizer):
-        with rasterizer.surface as canvas:
-            canvas.clipRRect(
-                skia.RRect.MakeRectXY(
-                    skia.Rect.MakeLTRB(
-                        self.rect.left(), self.rect.top() - scroll,
-                        self.rect.right(), self.rect.bottom() - scroll),
-                    self.radius, self.radius))
+class TextNode:
+    def __init__(self, text):
+        self.text = text
+        self.tag = None
+        self.children = []
 
-class DrawImage:
-    def __init__(self, image, rect):
-        self.image = image
-        self.rect = rect
+    def __repr__(self):
+        return self.text.replace("\n", "\\n")
+        
+def parse(tokens):
+    currently_open = []
+    for tok in tokens:
+        implicit_tags(tok, currently_open)
+        if isinstance(tok, Text):
+            node = TextNode(tok.text)
+            if not currently_open: continue
+            node.parent = currently_open[-1]
+            currently_open[-1].children.append(node)
+        elif tok.tag.startswith("/"):
+            node = currently_open.pop()
+            if not currently_open: return node
+            currently_open[-1].children.append(node)
+        elif tok.tag in SELF_CLOSING_TAGS:
+            node = ElementNode(tok.tag, tok.attributes)
+            node.parent = currently_open[-1]
+            currently_open[-1].children.append(node)
+        elif tok.tag.startswith("!"):
+            continue
+        else:
+            node = ElementNode(tok.tag, tok.attributes)
+            node.parent = currently_open[-1]
+            currently_open.append(node)
+    while currently_open:
+        node = currently_open.pop()
+        if not currently_open: return node
+        currently_open[-1].children.append(node)
 
-    def execute(self, scroll, rasterizer):
-        with rasterizer.surface as canvas:
-            canvas.drawImage(
-                self.image, self.rect.left(),
-                self.rect.top() - scroll)
+HEAD_TAGS = [
+    "base", "basefont", "bgsound", "noscript",
+    "link", "meta", "title", "style", "script",
+]
+            
+def implicit_tags(tok, currently_open):
+    tag = tok.tag if isinstance(tok, Tag) else None
+    while True:
+        open_tags = [node.tag for node in currently_open]
+        if open_tags == [] and tag != "html":
+            node = ElementNode("html", {})
+            node.parent = None
+            currently_open.append(node)
+        elif open_tags == ["html"] and tag not in ["head", "body", "/html"]:
+            if tag in HEAD_TAGS:
+                implicit = "head"
+            else:
+                implicit = "body"
+            node = ElementNode(implicit, {})
+            node.parent = currently_open[-1]
+            currently_open.append(node)
+        elif open_tags == ["html", "head"] and tag not in ["/head"] + HEAD_TAGS:
+            node = currently_open.pop()
+            currently_open[-1].children.append(node)
+        else:
+            break
 
-INPUT_WIDTH_PX = 200
+class CSSParser:
+    def __init__(self, s):
+        self.s = s
+
+    def whitespace(self, i):
+        while i < len(self.s) and self.s[i].isspace():
+            i += 1
+        return None, i
+
+    def literal(self, i, literal):
+        l = len(literal)
+        assert self.s[i:i+l] == literal
+        return None, i + l
+
+    def word(self, i):
+        j = i
+        while j < len(self.s) and self.s[j].isalnum() or self.s[j] in "-.":
+            j += 1
+        assert j > i
+        return self.s[i:j], j
+
+    def pair(self, i):
+        prop, i = self.word(i)
+        _, i = self.whitespace(i)
+        _, i = self.literal(i, ":")
+        _, i = self.whitespace(i)
+        val, i = self.word(i)
+        return (prop.lower(), val), i
+
+    def ignore_until(self, i, chars):
+        while i < len(self.s) and self.s[i] not in chars:
+            i += 1
+        return None, i
+
+    def body(self, i):
+        pairs = {}
+        _, i = self.literal(i, "{")
+        _, i = self.whitespace(i)
+        while i < len(self.s) and self.s[i] != "}":
+            try:
+                (prop, val), i = self.pair(i)
+                pairs[prop] = val
+                _, i = self.whitespace(i)
+                _, i = self.literal(i, ";")
+            except AssertionError:
+                _, i = self.ignore_until(i, [";", "}"])
+                if i < len(self.s) and self.s[i] == ";":
+                    _, i = self.literal(i, ";")
+            _, i = self.whitespace(i)
+        _, i = self.literal(i, "}")
+        return pairs, i
+
+    def selector(self, i):
+        if self.s[i] == "#":
+            _, i = self.literal(i, "#")
+            name, i = self.word(i)
+            return IdSelector(name), i
+        elif self.s[i] == ".":
+            _, i = self.literal(i, ".")
+            name, i = self.word(i)
+            return ClassSelector(name), i
+        else:
+            name, i = self.word(i)
+            return TagSelector(name.lower()), i
+
+    def rule(self, i):
+        selector, i = self.selector(i)
+        _, i = self.whitespace(i)
+        body, i = self.body(i)
+        return (selector, body), i
+
+    def file(self, i):
+        rules = []
+        _, i = self.whitespace(i)
+        while i < len(self.s):
+            try:
+                rule, i = self.rule(i)
+                rules.append(rule)
+            except AssertionError:
+                _, i = self.ignore_until(i, "}")
+                _, i = self.literal(i, "}")
+            _, i = self.whitespace(i)
+        return rules, i
+
+    def parse(self):
+        rules, _ = self.file(0)
+        return rules
+
+class TagSelector:
+    def __init__(self, tag):
+        self.tag = tag
+
+    def matches(self, node):
+        return self.tag == node.tag
+
+    def priority(self):
+        return 1
+
+class ClassSelector:
+    def __init__(self, cls):
+        self.cls = cls
+
+    def matches(self, node):
+        return self.cls in node.attributes.get("class", "").split()
+
+    def priority(self):
+        return 16
+
+class IdSelector:
+    def __init__(self, id):
+        self.id = id
+
+    def matches(self, node):
+        return self.id == node.attributes.get("id", "")
+
+    def priority(self):
+        return 256
+
+INHERITED_PROPERTIES = {
+    "font-style": "normal",
+    "font-weight": "normal",
+    "font-size": "16px",
+    "color": "black",
+}
+
+def style(node, parent, rules):
+    if isinstance(node, TextNode):
+        node.style = parent.style
+    else:
+        for selector, pairs in rules:
+            if selector.matches(node):
+                for property in pairs:
+                    if property not in node.style:
+                        node.style[property] = pairs[property]
+        for property, default in INHERITED_PROPERTIES.items():
+            if property not in node.style:
+                if parent:
+                    node.style[property] = parent.style[property]
+                else:
+                    node.style[property] = default
+    for child in node.children:
+        style(child, node, rules)
+
+WIDTH, HEIGHT = 800, 600
+HSTEP, VSTEP = 13, 18
+
+SCROLL_STEP = 100
 
 class LineLayout:
-    def __init__(self, node, parent, previous):
+    def __init__(self, node, parent):
         self.node = node
         self.parent = parent
-        self.previous = previous
+        self.cx = 0
+        self.laid_out = False
         self.children = []
-        self.x = None
-        self.y = None
-        self.width = None
-        self.height = None
 
-    def layout(self):
-        self.width = self.parent.width
-        self.x = self.parent.x
+    def append(self, child):
+        self.children.append(child)
+        child.parent = self
+        self.cx += child.w + child.font.measure(" ")
 
-        if self.previous:
-            self.y = self.previous.y + self.previous.height
-        else:
-            self.y = self.parent.y
+    def size(self):
+        self.w = self.parent.w
+        self.compute_height()
 
-        for word in self.children:
-            word.layout()
-
+    def compute_height(self):
         if not self.children:
-            self.height = 0
+            self.h = 0
+            self.max_ascent = 0
+            self.max_descent = 0
+            self.metrics = None
+            self.cxs = []
             return
+        self.metrics = [child.font.metrics() for child in self.children]
+        self.max_ascent = max([metric["ascent"] for metric in self.metrics])
+        self.max_descent = max([metric["descent"] for metric in self.metrics])
+        self.h = 1.25 * (self.max_descent + self.max_ascent)
 
-        max_ascent = max([-word.font.getMetrics().fAscent 
-                          for word in self.children])
-        baseline = self.y + 1.25 * max_ascent
-        for word in self.children:
-            word.y = baseline + word.font.getMetrics().fAscent
-        max_descent = max([word.font.getMetrics().fDescent
-                           for word in self.children])
-        self.height = 1.25 * (max_ascent + max_descent)
-
-    def paint(self, display_list):
+        cx = 0
+        self.cxs = []
         for child in self.children:
-            child.paint(display_list)
+            self.cxs.append(cx)
+            cx += child.w + child.font.measure(" ")
 
-    def __repr__(self):
-        return "LineLayout(x={}, y={}, width={}, height={})".format(
-            self.x, self.y, self.width, self.height)
+    def position(self):
+        baseline = self.y + 1.25 * self.max_ascent
+        if self.children:
+            for cx, child, metrics in \
+              zip(self.cxs, self.children, self.metrics):
+                child.x = self.x + cx
+                child.y = baseline - metrics["ascent"]
 
-def font_style(weight, style):
-    skia_weight = skia.FontStyle.kNormal_Weight
-    if weight == "bold":
-        skia_weight = skia.FontStyle.kBold_Weight
-    skia_style = skia.FontStyle.kUpright_Slant
-    if style == "italic":
-        skia_style = skia.FontStyle.kItalic_Slant
-    return skia.FontStyle(skia_weight, skia.FontStyle.kNormal_Width, skia_style)
+    def paint(self, to):
+        for child in self.children:
+            child.paint(to)
+
+FONT_CACHE = {}
+
+def GetFont(size, weight, style):
+    key = (size, weight, style)
+    value = FONT_CACHE.get(key)
+    if value: return value
+    value = tkinter.font.Font(size=size, weight=weight, slant=style)
+    FONT_CACHE[key] = value
+    return value
 
 class TextLayout:
-    def __init__(self, node, word, parent, previous):
+    def __init__(self, node, word):
         self.node = node
-        self.word = word
         self.children = []
-        self.parent = parent
-        self.previous = previous
-        self.x = None
-        self.y = None
-        self.width = None
-        self.height = None
-        self.font = None
+        self.word = word
+        self.display_item = None
 
-    def layout(self):
+    def size(self):
+        self.display_item = None
+
         weight = self.node.style["font-weight"]
         style = self.node.style["font-style"]
         if style == "normal": style = "roman"
-        size = int(self.node.style["font-size"][:-2])
-        self.font = skia.Font(
-            skia.Typeface('Arial', font_style(weight, style)), size)
+        size = int(px(self.node.style["font-size"]) * .75)
+        self.font = GetFont(size, weight, style) 
 
-        # Do not set self.y!!!
-        self.width = self.font.measureText(self.word)
+        self.w = self.font.measure(self.word)
+        self.compute_height()
 
-        if self.previous:
-            space = self.previous.font.measureText(" ")
-            self.x = self.previous.x + space + self.previous.width
-        else:
-            self.x = self.parent.x
+    def compute_height(self):
+        self.h = self.font.metrics('linespace')
 
-        self.height = linespace(self.font)
+    def position(self):
+        pass
 
-    def paint(self, display_list):
-        color = self.node.style["color"]
-        display_list.append(
-            DrawText(self.x, self.y, self.word, self.font, color))
-    
-    def __repr__(self):
-        return "TextLayout(x={}, y={}, width={}, height={}".format(
-            self.x, self.y, self.width, self.height)
+    def paint(self, to):
+        if not self.display_item:
+            color = self.node.style["color"]
+            self.display_item = DrawText(self.x, self.y, self.word, self.font, color)
+        to.append(self.display_item)
 
 class InputLayout:
-    def __init__(self, node, parent, previous):
+    def __init__(self, node):
         self.node = node
         self.children = []
-        self.parent = parent
-        self.previous = previous
-        self.x = None
-        self.y = None
-        self.width = None
-        self.height = None
-        self.font = None
 
-    def layout(self):
+    def size(self):
         weight = self.node.style["font-weight"]
         style = self.node.style["font-style"]
         if style == "normal": style = "roman"
-        size = int(self.node.style["font-size"][:-2])
-        self.font = skia.Font(
-            skia.Typeface('Arial', font_style(weight, style)), size)
+        size = int(px(self.node.style["font-size"]) * .75)
+        self.font = tkinter.font.Font(size=size, weight=weight, slant=style)
+        self.w = 200
+        self.compute_height()
 
-        self.width = style_length(self.node, "width", INPUT_WIDTH_PX)
-        self.height = style_length(
-            self.node, "height", linespace(self.font))
+    def compute_height(self):
+        self.h = 20
 
-        if self.previous:
-            space = self.previous.font.measureText(" ")
-            self.x = self.previous.x + space + self.previous.width
-        else:
-            self.x = self.parent.x
+    def position(self):
+        pass
 
-    def paint(self, display_list):
-        (paint_x, paint_y) = paint_coords(self.node, self.x, self.y)
-
-        rect = skia.Rect.MakeLTRB(
-            paint_x, paint_y, paint_x + self.width,
-            paint_y + self.height)
-
-        restore_count = paint_visual_effects(self.node, display_list, rect)
-
-        paint_background(self.node, display_list, rect)
+    def paint(self, to):
+        x1, x2 = self.x, self.x + self.w
+        y1, y2 = self.y, self.y + self.h
+        bgcolor = "light gray" if self.node.tag == "input" else "yellow"
+        to.append(DrawRect(x1, y1, x2, y2, bgcolor))
 
         if self.node.tag == "input":
             text = self.node.attributes.get("value", "")
-        elif self.node.tag == "button":
-            text = self.node.children[0].text
-
-        color = self.node.style["color"]
-        display_list.append(
-            DrawText(self.x, self.y, text, self.font, color))
-
-        restore_count = restore_count + \
-            paint_clip_path(self.node, display_list, rect)
-
-        for i in range(0, restore_count):
-            display_list.append(Restore(rect))
-
-    def __repr__(self):
-        return "InputLayout(x={}, y={}, width={}, height={})".format(
-            self.x, self.y, self.width, self.height)
-
-def style_length(node, style_name, default_value):
-    style_val = node.style.get(style_name)
-    if style_val:
-        return int(style_val[:-2])
-    else:
-        return default_value
-
-def paint_clip_path(node, display_list, rect):
-    clip_path = node.style.get("clip-path")
-    if clip_path:
-        percent = parse_clip_path(clip_path)
-        if percent:
-            width = rect.right() - rect.left()
-            height = rect.bottom() - rect.top()
-            reference_val = \
-                math.sqrt(width * width + 
-                    height * height) / math.sqrt(2)
-            radius = reference_val * percent / 100
-            (center_x, center_y) = center_point(rect)
-            display_list.append(CircleMask(
-                center_x, center_y, radius, rect))
-            return 1
-    return 0
-
-def paint_visual_effects(node, display_list, rect):
-    restore_count = 0
-    
-    transform_str = node.style.get("transform", "")
-    if transform_str:
-        display_list.append(Save(rect))
-        restore_count = restore_count + 1
-        degrees = parse_rotation_transform(transform_str)
-        display_list.append(Rotate(degrees, rect))
-
-    blend_mode_str = node.style.get("mix-blend-mode")
-    blend_mode = skia.BlendMode.kSrcOver
-    if blend_mode_str:
-        blend_mode = parse_blend_mode(blend_mode_str)
-
-    opacity = float(node.style.get("opacity", "1.0"))
-    if opacity != 1.0 or blend_mode_str:
-        paint = skia.Paint(Alphaf=opacity, BlendMode=blend_mode)
-        display_list.append(SaveLayer(paint, rect))
-        restore_count = restore_count + 1
-
-    clip_path = node.style.get("clip-path")
-    if clip_path:
-        display_list.append(SaveLayer(skia.Paint(), rect))
-        restore_count = restore_count + 1
-
-    border_radius = node.style.get("border-radius")
-    if border_radius:
-        radius = int(border_radius[:-2])
-        display_list.append(Save(rect))
-        display_list.append(ClipRRect(rect, radius))
-        restore_count = restore_count + 1
-
-    return restore_count
-
-def paint_background(node, display_list, rect):
-    bgcolor = node.style.get("background-color",
-                             "transparent")
-    if bgcolor != "transparent":
-        display_list.append(DrawRect(rect, bgcolor))
-
-    background_image = node.style.get("background-image")
-    if background_image:
-        display_list.append(Save(rect))
-        display_list.append(ClipRect(rect))
-        display_list.append(DrawImage(node.background_image,
-            rect))
-        display_list.append(Restore(rect))
-
-def paint_coords(node, x, y):
-    if not node.style.get("position") == "relative":
-        return (x, y)
-
-    paint_x = x
-    paint_y = y
-
-    left = node.style.get("left")
-    if left:
-        paint_x = paint_x + int(left[:-2])
-    top = node.style.get("top")
-    if top:
-        paint_y = paint_y + int(top[:-2])
-
-    return (paint_x, paint_y)
-
-class BlockLayout:
-    def __init__(self, node, parent, previous):
-        self.node = node
-        self.parent = parent
-        self.previous = previous
-        self.children = []
-        self.x = None
-        self.y = None
-        self.width = None
-        self.height = None
-
-    def layout(self):
-        previous = None
-        for child in self.node.children:
-            if layout_mode(child) == "inline":
-                next = InlineLayout(child, self, previous)
-            else:
-                next = BlockLayout(child, self, previous)
-            self.children.append(next)
-            previous = next
-
-        self.width = style_length(
-            self.node, "width", self.parent.width)
-        self.x = self.parent.x
-
-        if self.previous:
-            self.y = self.previous.y + self.previous.height
         else:
-            self.y = self.parent.y
-
-        for child in self.children:
-            child.layout()
-
-        self.height = style_length(
-            self.node, "height",
-            sum([child.height for child in self.children]))
-
-
-    def paint(self, display_list):
-        (paint_x, paint_y) = paint_coords(self.node, self.x, self.y)
-        rect = skia.Rect.MakeLTRB(
-            paint_x, paint_y,
-            paint_x + self.width, paint_y + self.height)
-
-        restore_count = paint_visual_effects(
-            self.node, display_list, rect)
-
-        paint_background(self.node, display_list, rect)
-
-        for child in self.children:
-            child.paint(display_list)
-
-        restore_count = restore_count + \
-            paint_clip_path(self.node, display_list, rect)
-
-        for i in range(0, restore_count):
-            display_list.append(Restore(rect))
-
-    def __repr__(self):
-        return "BlockLayout(x={}, y={}, width={}, height={})".format(
-            self.x, self.x, self.width, self.height)
+            text = self.node.children[0].text
+        color = self.node.style["color"]
+        to.append(DrawText(self.x, self.y, text, self.font, color))
 
 class InlineLayout:
-    def __init__(self, node, parent, previous):
+    def __init__(self, node, parent):
         self.node = node
         self.parent = parent
-        self.previous = previous
-        self.children = []
-        self.x = None
-        self.y = None
-        self.width = None
-        self.height = None
-        self.display_list = None
 
-    def layout(self):
-        self.width = style_length(
-            self.node, "width", self.parent.width)
+    def size(self):
+        self.children = [LineLayout(self.node, self)]
+        self.mt = self.bt = self.pt = 0
+        self.mr = self.br = self.pr = 0
+        self.mb = self.bb = self.pb = 0
+        self.ml = self.bl = self.pl = 0
 
-        self.x = self.parent.x
+        self.w = self.parent.w - self.parent.pl - self.parent.pr \
+            - self.parent.bl - self.parent.br
 
-        if self.previous:
-            self.y = self.previous.y + self.previous.height
-        else:
-            self.y = self.parent.y
-
-        self.new_line()
         self.recurse(self.node)
-        
-        for line in self.children:
-            line.layout()
+        self.flush()
+        self.children.pop()
+        self.compute_height()
 
-        self.height = style_length(
-            self.node, "height",
-            sum([line.height for line in self.children]))
+    def compute_height(self):
+        self.h = 0
+        for child in self.children:
+            self.h += child.h
 
     def recurse(self, node):
-        if isinstance(node, Text):
+        if isinstance(node, TextNode):
             self.text(node)
+        elif node.tag == "br":
+            self.flush()
+        elif node.tag == "input":
+            self.input(node)
+        elif node.tag == "button":
+            self.input(node)
         else:
-            if node.tag == "br":
-                self.new_line()
-            elif node.tag == "input" or node.tag == "button":
-                self.input(node)
-            else:
-                for child in node.children:
-                    self.recurse(child)
-
-    def new_line(self):
-        self.previous_word = None
-        self.cursor_x = self.x
-        last_line = self.children[-1] if self.children else None
-        new_line = LineLayout(self.node, self, last_line)
-        self.children.append(new_line)
-
-    def get_font(self, node):
-        weight = node.style["font-weight"]
-        style = node.style["font-style"]
-        size = int(node.style["font-size"][:-2])
-        return skia.Font(
-            skia.Typeface('Arial', font_style(weight, style)), size)
+            for child in node.children:
+                self.recurse(child)
 
     def text(self, node):
-        font = self.get_font(node)
         for word in node.text.split():
-            w = font.measureText(word)
-            if self.cursor_x + w > self.x + self.width:
-                self.new_line()
-            line = self.children[-1]
-            text = TextLayout(node, word, line, self.previous_word)
-            line.children.append(text)
-            self.previous_word = text
-            self.cursor_x += w + font.measureText(" ")
+            child = TextLayout(node, word)
+            child.size()
+            if self.children[-1].cx + child.w > self.w:
+                self.flush()
+            self.children[-1].append(child)
 
     def input(self, node):
-        w = INPUT_WIDTH_PX
-        if self.cursor_x + w > self.x + self.width:
-            self.new_line()
-        line = self.children[-1]
-        input = InputLayout(node, line, self.previous_word)
-        line.children.append(input)
-        self.previous_word = input
-        size = int(float(node.style["font-size"][:-2]) * .75)
-        font = self.get_font(node)
-        self.cursor_x += w + font.measureText(" ")
+        child = InputLayout(node)
+        child.size()
+        if self.children[-1].cx + child.w > self.w:
+            self.flush()
+        self.children[-1].append(child)
 
-    def paint(self, display_list):
-        (paint_x, paint_y) = paint_coords(self.node, self.x, self.y)
+    def flush(self):
+        child = self.children[-1]
+        child.size()
+        self.children.append(LineLayout(self.node, self))
 
-        rect = skia.Rect.MakeLTRB(
-            paint_x, paint_y, paint_x + self.width,
-            paint_y + self.height)
-
-        restore_count = paint_visual_effects(self.node, display_list, rect)
-
-        paint_background(self.node, display_list, rect)
-
+    def position(self):
+        cy = self.y
         for child in self.children:
-            child.paint(display_list)
+            child.x = self.x
+            child.y = cy
+            child.position()
+            cy += child.h
 
-        restore_count = restore_count + \
-            paint_clip_path(self.node, display_list, rect)
+    def paint(self, to):
+        for child in self.children:
+            child.paint(to)
 
-        for i in range(0, restore_count):
-            display_list.append(Restore(rect))
+def px(s):
+    if s.endswith("px"):
+        return int(s[:-2])
+    else:
+        return 0
 
-    def __repr__(self):
-        return "InlineLayout(x={}, y={}, width={}, height={})".format(
-            self.x, self.y, self.width, self.height)
+class BlockLayout:
+    def __init__(self, node, parent):
+        self.node = node
+        self.parent = parent
+
+        self.x = -1
+        self.y = -1
+        self.w = -1
+        self.h = -1
+
+    def has_block_children(self):
+        for child in self.node.children:
+            if isinstance(child, TextNode):
+                if not child.text.isspace():
+                    return False
+            elif child.style.get("display", "block") == "inline":
+                return False
+        return True
+
+    def size(self):
+        self.children = []
+        # block layout here
+        if self.has_block_children():
+            for child in self.node.children:
+                if isinstance(child, TextNode): continue
+                self.children.append(BlockLayout(child, self))
+        else:
+            self.children.append(InlineLayout(self.node, self))
+
+        self.mt = px(self.node.style.get("margin-top", "0px"))
+        self.bt = px(self.node.style.get("border-top-width", "0px"))
+        self.pt = px(self.node.style.get("padding-top", "0px"))
+        self.mr = px(self.node.style.get("margin-right", "0px"))
+        self.br = px(self.node.style.get("border-right-width", "0px"))
+        self.pr = px(self.node.style.get("padding-right", "0px"))
+        self.mb = px(self.node.style.get("margin-bottom", "0px"))
+        self.bb = px(self.node.style.get("border-bottom-width", "0px"))
+        self.pb = px(self.node.style.get("padding-bottom", "0px"))
+        self.ml = px(self.node.style.get("margin-left", "0px"))
+        self.bl = px(self.node.style.get("border-left-width", "0px"))
+        self.pl = px(self.node.style.get("padding-left", "0px"))
+
+        self.w = self.parent.w - self.parent.pl - self.parent.pr \
+            - self.parent.bl - self.parent.br \
+            - self.ml - self.mr
+        for child in self.children:
+            child.size()
+        self.compute_height()
+
+    def compute_height(self):
+        self.h = 0
+        for child in self.children:
+            self.h += child.mt + child.h + child.mb
+
+    def position(self):
+        self.y += self.mt
+        self.x += self.ml
+
+        y = self.y
+        for child in self.children:
+            child.x = self.x + self.pl + self.bl
+            child.y = y
+            child.position()
+            y += child.mt + child.h + child.mb
+
+    def paint(self, to):
+        if self.node.tag == "pre":
+            x2, y2 = self.x + self.w, self.y + self.h
+            to.append(DrawRect(self.x, self.y, x2, y2, "gray"))
+        for child in self.children:
+            child.paint(to)
 
 class DocumentLayout:
     def __init__(self, node):
         self.node = node
         self.parent = None
-        self.previous = None
         self.children = []
 
-    def layout(self):
-        child = BlockLayout(self.node, self, None)
+        self.x = -1
+        self.y = -1
+        self.w = -1
+        self.h = -1
+
+    def size(self):
+        child = BlockLayout(self.node, self)
         self.children.append(child)
 
-        self.width = WIDTH - 2*HSTEP
-        self.x = HSTEP
-        self.y = VSTEP
-        child.layout()
-        self.height = child.height + 2*VSTEP
+        self.w = WIDTH
+        self.mt = self.bt = self.pt = 0
+        self.mr = self.br = self.pr = 0
+        self.mb = self.bb = self.pb = 0
+        self.ml = self.bl = self.pl = 0
 
-    def paint(self, display_list):
-        self.children[0].paint(display_list)
+        child.size()
+        self.compute_height()
 
-    def __repr__(self):
-        return "DocumentLayout()"
+    def compute_height(self):
+        self.h = self.children[0].h
 
-class CSSParser:
-    def __init__(self, s):
-        self.s = s
-        self.i = 0
+    def position(self):
+        child = self.children[0]
+        child.x = self.x = 0
+        child.y = self.y = 0
+        child.position()
 
-    def whitespace(self):
-        while self.i < len(self.s) and self.s[self.i].isspace():
-            self.i += 1
+    def paint(self, to):
+        self.children[0].paint(to)
 
-    def literal(self, literal):
-        assert self.i < len(self.s) and self.s[self.i] == literal
-        self.i += 1
+class DrawText:
+    def __init__(self, x1, y1, text, font, color):
+        self.x1 = x1
+        self.y1 = y1
+        self.text = text
+        self.font = font
+        self.color = color
 
-    def word(self):
-        start = self.i
-        while self.i < len(self.s):
-            if self.s[self.i].isalnum() or self.s[self.i] in "#-.%()\"'":
-                self.i += 1
-            else:
-                break
-        assert self.i > start
-        return self.s[start:self.i]
+        self.y2 = y1 + font.measure("linespace")
 
-    def pair(self):
-        prop = self.word()
-        self.whitespace()
-        self.literal(":")
-        self.whitespace()
-        val = self.word()
-        return prop.lower(), val
+    def draw(self, scroll, canvas):
+        canvas.create_text(
+            self.x1, self.y1 - scroll,
+            text=self.text,
+            font=self.font,
+            fill=self.color,
+            anchor='nw',
+        )
 
-    def ignore_until(self, chars):
-        while self.i < len(self.s):
-            if self.s[self.i] in chars:
-                return self.s[self.i]
-            else:
-                self.i += 1
+class DrawRect:
+    def __init__(self, x1, y1, x2, y2, color):
+        self.x1 = x1
+        self.y1 = y1
+        self.x2 = x2
+        self.y2 = y2
+        self.color = color
 
-    def body(self):
-        pairs = {}
-        while self.i < len(self.s) and self.s[self.i] != "}":
-            try:
-                prop, val = self.pair()
-                pairs[prop.lower()] = val
-                self.whitespace()
-                self.literal(";")
-                self.whitespace()
-            except AssertionError:
-                why = self.ignore_until([";", "}"])
-                if why == ";":
-                    self.literal(";")
-                    self.whitespace()
-                else:
-                    break
-        return pairs
+    def draw(self, scroll, canvas):
+        canvas.create_rectangle(
+            self.x1, self.y1 - scroll,
+            self.x2, self.y2 - scroll,
+            width=0,
+            fill=self.color,
+        )
 
-    def selector(self):
-        out = TagSelector(self.word().lower())
-        self.whitespace()
-        while self.i < len(self.s) and self.s[self.i] != "{":
-            tag = self.word()
-            descendant = TagSelector(tag.lower())
-            out = DescendantSelector(out, descendant)
-            self.whitespace()
-        return out
-
-    def parse(self):
-        rules = []
-        while self.i < len(self.s):
-            try:
-                self.whitespace()
-                selector = self.selector()
-                self.literal("{")
-                self.whitespace()
-                body = self.body()
-                self.literal("}")
-                rules.append((selector, body))
-            except AssertionError:
-                why = self.ignore_until(["}"])
-                if why == "}":
-                    self.literal("}")
-                    self.whitespace()
-                else:
-                    break
-        return rules
-
-def parse_style_url(url_str):
-    return url_str[5:][:-2]
-
-def get_images(image_url_strs, base_url, images):
-    for image_url_str in image_url_strs:
-        image_url = parse_style_url(image_url_str)
-        header, body_bytes = request(
-            resolve_url(image_url, base_url),
-            headers={})
-        picture_stream = io.BytesIO(body_bytes)
-
-        pil_image = Image.open(picture_stream)
-        if pil_image.mode == "RGBA":
-            pil_image_bytes = pil_image.tobytes()
-        else:
-            pil_image_bytes = pil_image.convert("RGBA").tobytes()
-        images[image_url] = skia.Image.frombytes(
-            array=pil_image_bytes,
-            dimensions=pil_image.size,
-            colorType=skia.kRGBA_8888_ColorType)
-
-def style(node, rules, url, images):
-    node.style = {}
-    for property, default_value in INHERITED_PROPERTIES.items():
-        if node.parent:
-            node.style[property] = node.parent.style[property]
-        else:
-            node.style[property] = default_value
-    for selector, body in rules:
-        if not selector.matches(node): continue
-        for property, value in body.items():
-            computed_value = compute_style(node, property, value)
-            if not computed_value: continue
-            node.style[property] = computed_value
-    if isinstance(node, Element) and "style" in node.attributes:
-        pairs = CSSParser(node.attributes["style"]).body()
-        image_url_strs = []
-        for property, value in pairs.items():
-            computed_value = compute_style(node, property, value)
-            node.style[property] = computed_value
-            if property == 'background-image':
-                image_url_strs.append(value)
-        get_images(image_url_strs, url, images)
-    if node.style.get('background-image'):
-        node.background_image = \
-            images[parse_style_url(
-                node.style.get('background-image'))]
+def find_links(node, lst):
+    if not isinstance(node, ElementNode): return
+    if node.tag == "link" and \
+       node.attributes.get("rel", "") == "stylesheet" and \
+       "href" in node.attributes:
+        lst.append(node.attributes["href"])
     for child in node.children:
-        style(child, rules, url, images)
+        find_links(child, lst)
+    return lst
 
-SCROLL_STEP = 100
-CHROME_PX = 100
+def resolve_url(url, current):
+    if "://" in url:
+        return url
+    elif url.startswith("/"):
+        return "/".join(current.split("/")[:3]) + url
+    else:
+        return current.rsplit("/", 1)[0] + "/" + url
 
-class Tab:
+def find_layout(x, y, tree):
+    for child in reversed(tree.children):
+        result = find_layout(x, y, child)
+        if result: return result
+    if tree.x <= x < tree.x + tree.w and \
+       tree.y <= y < tree.y + tree.h:
+        return tree
+
+def find_inputs(elt, out):
+    if not isinstance(elt, ElementNode): return
+    if elt.tag == "input" and "name" in elt.attributes:
+        out.append(elt)
+    for child in elt.children:
+        find_inputs(child, out)
+    return out
+
+def find_scripts(node, out):
+    if not isinstance(node, ElementNode): return
+    if node.tag == "script" and \
+       "src" in node.attributes:
+        out.append(node.attributes["src"])
+    for child in node.children:
+        find_scripts(child, out)
+    return out
+
+def find_selected(node, sel, out):
+    if not isinstance(node, ElementNode): return
+    if sel.matches(node):
+        out.append(node)
+    for child in node.children:
+        find_selected(child, sel, out)
+    return out
+
+def layout_for_node(tree, node):
+    if tree.node == node:
+        return tree
+    for child in tree.children:
+        out = layout_for_node(child, node)
+        if out: return out
+
+def is_link(node):
+    return isinstance(node, ElementNode) \
+        and node.tag == "a" and "href" in node.attributes
+
+def drawHTMLTree(node, indent=0):
+    print(" "*indent, type(node).__name__, " ", node, sep="")
+    for child in node.children:
+        drawHTMLTree(child, indent + 2)
+
+def drawLayoutTree(node, indent=0):
+    print(" "*indent, type(node).__name__, " ", node.node, sep="")
+    for child in node.children:
+        drawLayoutTree(child, indent + 2)
+
+REFRESH_RATE_MS = 16 # 16ms
+
+class Task:
+    def __init__(self, task_code, arg1=None, arg2=None):
+        self.task_code = task_code
+        self.arg1 = arg1
+        self.arg2 = arg2
+        self.__name__ = "task"
+
+    def __call__(self):
+        if self.arg2:
+            self.task_code(self.arg1, self.arg2)
+        elif self.arg1:
+            self.task_code(self.arg1)
+        else:
+            self.task_code()
+        # Prevent it accidentally running twice.
+        self.task_code = None
+        self.arg1 = None
+        self.arg2 = None
+
+class TaskQueue:
+    def __init__(self, lock):
+        self.tasks = []
+        self.lock = lock
+
+    def add_task(self, task_code):
+        self.lock.acquire(blocking=True)
+        self.tasks.append(task_code)
+        self.lock.release()
+
+    def has_tasks(self):
+        self.lock.acquire(blocking=True)
+        retval = len(self.tasks) > 0
+        self.lock.release()
+        return retval
+
+    def get_next_task(self):
+        self.lock.acquire(blocking=True)
+        retval = self.tasks.pop(0)
+        self.lock.release()
+        return retval
+
+class MainThreadRunner:
+    def __init__(self, browser):
+        self.lock = threading.Lock()
+        self.browser = browser
+        self.needs_animation_frame = False
+        self.main_thread = threading.Thread(target=self.run, args=())
+        self.script_tasks = TaskQueue(self.lock)
+        self.browser_tasks = TaskQueue(self.lock)
+
+    def schedule_animation_frame(self):
+        self.lock.acquire(blocking=True)
+        self.needs_animation_frame = True
+        self.lock.release()
+
+    def schedule_script_task(self, script):
+        self.script_tasks.add_task(script)
+
+    def schedule_browser_task(self, callback):
+        self.browser_tasks.add_task(callback)
+
+    def schedule_event_handler():
+        pass
+
+    def start(self):
+        self.main_thread.start()
+
+    def run(self):
+        while True:
+            self.lock.acquire(blocking=True)
+            needs_animation_frame = self.needs_animation_frame
+            self.lock.release()
+            if needs_animation_frame:
+                browser.run_animation_frame()
+                self.browser.commit()
+
+            browser_method = None
+            if self.browser_tasks.has_tasks():
+                browser_method = self.browser_tasks.get_next_task()
+            if browser_method:
+                browser_method()
+
+            script = None
+            if self.script_tasks.has_tasks():
+                script = self.script_tasks.get_next_task()
+
+            if script:
+                script()
+
+            time.sleep(0.001) # 1ms
+
+class Browser:
     def __init__(self):
-        self.history = []
-        self.focus = None
+        self.window = tkinter.Tk()
+        self.canvas = tkinter.Canvas(
+            self.window,
+            width=WIDTH,
+            height=HEIGHT
+        )
+        self.canvas.pack()
         self.cookies = {}
 
-        with open("browser8.css") as f:
-            self.default_style_sheet = CSSParser(f.read()).parse()
-
-    def cookie_string(self):
-        origin = url_origin(self.history[-1])
-        cookie_string = ""
-        if not origin in self.cookies:
-            return cookie_string
-        for key, value in self.cookies[origin].items():
-            cookie_string += "&" + key + "=" + value
-        return cookie_string[1:]
-
-    def load(self, url, body=None):
-        self.scroll = 0
-        self.url = url
-        self.history.append(url)
-        req_headers = { "Cookie": self.cookie_string() }        
-        headers, body = request(url, headers=req_headers, payload=body)
-        if "set-cookie" in headers:
-            if ";" in headers["set-cookie"]:
-                kv, params = headers["set-cookie"].split(";", 1)
-            else:
-                kv = headers["set-cookie"]
-            key, value = kv.split("=", 1)
-            origin = url_origin(self.history[-1])
-            self.cookies.setdefault(origin, {})[key] = value
-
-        self.nodes = HTMLParser(body).parse()
-
-        self.js = JSContext(self)
-        scripts = [node.attributes["src"] for node
-                   in tree_to_list(self.nodes, [])
-                   if isinstance(node, Element)
-                   and node.tag == "script"
-                   and "src" in node.attributes]
-        for script in scripts:
-            header, body = request(resolve_url(script, url),
-                headers=req_headers)
-            try:
-                print("Script returned: ", self.js.run(body))
-            except dukpy.JSRuntimeError as e:
-                print("Script", script, "crashed", e)
-
-        self.rules = self.default_style_sheet.copy()
-        links = [node.attributes["href"]
-                 for node in tree_to_list(self.nodes, [])
-                 if isinstance(node, Element)
-                 and node.tag == "link"
-                 and "href" in node.attributes
-                 and node.attributes.get("rel") == "stylesheet"]
-        for link in links:
-            try:
-                header, body = request(resolve_url(link, url),
-                    headers=req_headers)
-            except:
-                continue
-            self.rules.extend(CSSParser(body).parse())
-
-        image_url_strs = [rule[1]['background-image']
-                 for rule in self.rules
-                 if 'background-image' in rule[1]]
-
-        self.images = {}
-        get_images(image_url_strs, url, self.images)
-
-        self.render()
-
-    def render(self):
-        style(self.nodes, sorted(self.rules, key=cascade_priority),
-            self.url, self.images)
-        self.document = DocumentLayout(self.nodes)
-        self.document.layout()
-        self.display_list = []
-        self.document.paint(self.display_list)
-
-    def draw(self, rasterizer):
-        for cmd in self.display_list:
-            if cmd.rect.top() > self.scroll + HEIGHT - CHROME_PX: continue
-            if cmd.rect.bottom() < self.scroll: continue
-            cmd.execute(self.scroll - CHROME_PX, rasterizer)
-
-        if self.focus:
-            obj = [obj for obj in tree_to_list(self.document, [])
-                   if obj.node == self.focus][0]
-            text = self.focus.attributes.get("value", "")
-            x = obj.x + obj.font.measureText(text)
-            y = obj.y - self.scroll + CHROME_PX
-            rasterizer.draw_polyline(x, y, x, y + obj.height)
-
-    def scrolldown(self):
-        max_y = self.document.height - HEIGHT
-        self.scroll = min(self.scroll + SCROLL_STEP, max_y)
-
-    def click(self, x, y):
+        self.history = []
         self.focus = None
-        y += self.scroll
-        objs = [obj for obj in tree_to_list(self.document, [])
-                if obj.x <= x < obj.x + obj.width
-                and obj.y <= y < obj.y + obj.height]
-        if not objs: return
-        elt = objs[-1].node
-        if elt and self.js.dispatch_event("click", elt): return
+        self.address_bar = ""
+        self.scroll = 0
+        self.display_list = []
+
+        self.draw_display_list = []
+        self.needs_draw = False
+
+        self.document = None
+
+        self.main_thread_timer = Timer()
+        self.compositor_thread_timer = Timer()
+        self.window.bind("<Down>", self.scrolldown)
+        self.window.bind("<Button-1>", self.compositor_handle_click)
+        self.window.bind("<Key>", self.compositor_keypress)
+        self.window.bind("<Return>", self.press_enter)
+
+        self.reflow_roots = []
+        self.needs_layout_tree_rebuild = False
+        self.needs_animation_frame = False
+        self.display_scheduled = False
+        self.needs_raf_callbacks = False
+
+        self.frame_count = 0
+        self.compositor_lock = threading.Lock()
+
+        self.needs_quit = False
+
+
+    def commit(self):
+        self.compositor_lock.acquire(blocking=True)
+        self.needs_draw = True
+        self.draw_display_list = self.display_list.copy()
+        self.compositor_lock.release()
+
+    def start(self):
+        self.main_thread_runner = MainThreadRunner(self)
+        self.main_thread_runner.start()
+        self.canvas.after(1, self.maybe_draw)
+
+    def maybe_draw(self):
+        self.compositor_lock.acquire(blocking=True)
+        if self.needs_quit:
+            sys.exit()
+        if self.needs_animation_frame and not self.display_scheduled:
+            self.canvas.after(
+                REFRESH_RATE_MS,
+                self.main_thread_runner.schedule_animation_frame)
+            self.display_scheduled = True
+
+        if self.needs_draw:
+            self.draw()
+        self.needs_draw = False
+        self.compositor_lock.release()
+        self.canvas.after(1, self.maybe_draw)
+
+    # Runs on the compositor thread
+    def compositor_handle_click(self, e):
+        self.focus = None
+        if e.y < 60:
+            # Browser chrome clicks can be handled without the main thread...
+            if 10 <= e.x < 35 and 10 <= e.y < 50:
+                self.go_back()
+            elif 50 <= e.x < 790 and 10 <= e.y < 50:
+                self.focus = "address bar"
+                self.address_bar = ""
+                self.set_needs_animation_frame()
+        else:
+            # ...but not clicks within the web page contents area
+            self.main_thread_runner.schedule_browser_task(
+                Task(self.handle_click, e))
+
+    # Runs on the main thread
+    def handle_click(self, e):
+        # Lock to check scroll, which is updated on the compositor thread.
+        self.compositor_lock.acquire(blocking=True)
+        x, y = e.x, e.y + self.scroll - 60
+        self.compositor_lock.release()
+        self.run_rendering_pipeline()
+        obj = find_layout(x, y, self.document)
+        if not obj: return
+        elt = obj.node
+        if elt and self.dispatch_event("click", elt): return
         while elt:
-            if isinstance(elt, Text):
+            if isinstance(elt, TextNode):
                 pass
-            elif elt.tag == "a" and "href" in elt.attributes:
+            elif is_link(elt):
                 url = resolve_url(elt.attributes["href"], self.url)
-                return self.load(url)
+                return self.schedule_load(url)
             elif elt.tag == "input":
                 elt.attributes["value"] = ""
-                self.focus = elt
-                return
+                self.focus = obj
+                self.set_needs_reflow(self.focus)
             elif elt.tag == "button":
-                while elt:
-                    if elt.tag == "form" and "action" in elt.attributes:
-                        return self.submit_form(elt)
-                    elt = elt.parent
+                self.submit_form(elt)
             elt = elt.parent
 
-    def submit_form(self, elt):
-        if self.js.dispatch_event("submit", elt): return
-        inputs = [node for node in tree_to_list(elt, [])
-                  if isinstance(node, Element)
-                  and node.tag == "input"
-                  and "name" in node.attributes]
+    # Runs on the compositor thread
+    def compositor_keypress(self, e):
+        if len(e.char) == 0: return
+        if not (0x20 <= ord(e.char) < 0x7f): return
 
+        if not self.focus:
+            return
+        elif self.focus == "address bar":
+            self.address_bar += e.char
+            self.set_needs_animation_frame()
+        else:
+            self.main_thread_runner.schedule_browser_task(
+                Task(self.keypress, e))
+
+    # Runs on the main thread
+    def keypress(self, e):
+        self.focus.node.attributes["value"] += e.char
+        self.dispatch_event("change", self.focus.node)
+        self.set_needs_reflow(self.focus)
+
+    def submit_form(self, elt):
+        while elt and elt.tag != "form":
+            elt = elt.parent
+        if not elt: return
+        if self.dispatch_event("submit", elt): return
+        inputs = find_inputs(elt, [])
         body = ""
         for input in inputs:
             name = input.attributes["name"]
             value = input.attributes.get("value", "")
-            name = urllib.parse.quote(name)
-            value = urllib.parse.quote(value)
-            body += "&" + name + "=" + value
-        body = body [1:]
-
+            body += "&" + name + "=" + value.replace(" ", "%20")
+        body = body[1:]
         url = resolve_url(elt.attributes["action"], self.url)
-        self.load(url, body)
+        self.schedule_load(url, body)
 
-    def keypress(self, char):
-        if self.focus:
-            if self.js.dispatch_event("keydown", self.focus): return
-            self.focus.attributes["value"] += char
-        self.document.paint(self.display_list) # TODO: is this necessary?
+    # Runs on the compositor thread
+    def press_enter(self, e):
+        if self.focus == "address bar":
+            self.focus = None
+            self.schedule_load(self.address_bar)
 
     def go_back(self):
         if len(self.history) > 1:
             self.history.pop()
             back = self.history.pop()
-            self.load(back)
+            self.schedule_load(back)
 
-WIDTH, HEIGHT = 800, 600
-HSTEP, VSTEP = 13, 18
+    def cookie_string(self):
+        cookie_string = ""
+        for key, value in self.cookies.items():
+            cookie_string += "&" + key + "=" + value
+        return cookie_string[1:]
 
-class Browser:
-    def __init__(self, sdl_window):
-        self.sdl_window = sdl_window
-        self.window_surface = SDL_GetWindowSurface(
-            self.sdl_window)
-        self.skia_surface = skia.Surface(WIDTH, HEIGHT)
+    # Runs on the compositor thread
+    def schedule_load(self, url, body=None):
+        self.main_thread_runner.schedule_browser_task(
+            Task(self.load, url, body))
 
-        self.tabs = []
-        self.active_tab = None
-        self.focus = None
-        self.address_bar = ""
+    # Runs on the main thread
+    def load(self, url, body=None):
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.start("Downloading")
+        self.address_bar = url
+        self.url = url
+        self.history.append(url)
+        req_headers = { "Cookie": self.cookie_string() }
+        headers, body = request(url, headers=req_headers, payload=body)
+        if "set-cookie" in headers:
+            kv, params = headers["set-cookie"].split(";", 1)
+            key, value = kv.split("=", 1)
+            origin = url_origin(self.history[-1])
+            self.cookies.setdefault(origin, {})[key] = value
 
-    def to_sdl_surface(skia_bytes):
-        depth = 32 # 4 bytes per pixel
-        pitch = 4 * WIDTH # 4 * WIDTH pixels per line on-screen
-        # Skia uses an ARGB format - alpha first byte, then
-        # through to blue as the last byte.
-        alpha_mask = 0xff000000
-        red_mask = 0x00ff0000
-        green_mask = 0x0000ff00
-        blue_mask = 0x000000ff
-        return SDL_CreateRGBSurfaceFrom(
-            skia_bytes, WIDTH, HEIGHT, depth, pitch,
-            red_mask, green_mask, blue_mask, alpha_mask)
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.start("Parsing HTML")
+        self.nodes = parse(lex(body))
+#        drawHTMLTree(self.nodes)
+        
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.start("Parsing CSS")
+        with open("browser8.css") as f:
+            self.rules = CSSParser(f.read()).parse()
 
-    def handle_down(self):
-        self.tabs[self.active_tab].scrolldown()
-        self.draw()
+        for link in find_links(self.nodes, []):
+            header, body = request(resolve_url(link, url), headers=req_headers)
+            self.rules.extend(CSSParser(body).parse())
 
-    def handle_click(self, e):
-        if e.y < CHROME_PX:
-            self.focus = None
-            if 40 <= e.x < 40 + 80 * len(self.tabs) and 0 <= e.y < 40:
-                self.active_tab = int((e.x - 40) / 80)
-            elif 10 <= e.x < 30 and 10 <= e.y < 30:
-                self.load("https://browser.engineering/")
-            elif 10 <= e.x < 35 and 40 <= e.y < 90:
-                self.tabs[self.active_tab].go_back()
-            elif 50 <= e.x < WIDTH - 10 and 40 <= e.y < 90:
-                self.focus = "address bar"
-                self.address_bar = ""
+        self.rules.sort(key=lambda x: x[0].priority())
+        self.rules.reverse()
+
+        self.run_scripts()
+        self.set_needs_layout_tree_rebuild()
+
+    def load_scripts(self, scripts):
+        req_headers = { "Cookie": self.cookie_string() }
+        for script in find_scripts(self.nodes, []):
+            header, body = request(
+                resolve_url(script, self.history[-1]), headers=req_headers)
+            scripts.append([header, body])
+
+    def script_run_wrapper(self, script_text):
+        return Task(self.js.evaljs, script_text)
+
+    def run_scripts(self):
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.start("Running JS")
+        self.setup_js()
+
+        scripts=[]
+        self.load_scripts(scripts)
+        for [header, body] in scripts:
+            self.main_thread_runner.schedule_script_task(
+                self.script_run_wrapper(body))
+
+    def setup_js(self):
+        self.js = dukpy.JSInterpreter()
+        self.node_to_handle = {}
+        self.handle_to_node = {}
+        self.js.export_function("log", print)
+        self.js.export_function("querySelectorAll", self.js_querySelectorAll)
+        self.js.export_function("getAttribute", self.js_getAttribute)
+        self.js.export_function("innerHTML", self.js_innerHTML)
+        self.js.export_function("cookie", self.cookie_string)
+        self.js.export_function(
+            "requestAnimationFrame",
+            self.js_requestAnimationFrame)
+        self.js.export_function("now", self.js_now)
+        with open("runtime13.js") as f:
+            self.main_thread_runner.schedule_script_task(
+                self.script_run_wrapper(f.read()))
+
+
+    def js_querySelectorAll(self, sel):
+        selector, _ = CSSParser(sel + "{").selector(0)
+        elts = find_selected(self.nodes, selector, [])
+        return [self.make_handle(elt) for elt in elts]
+    
+    def js_getAttribute(self, handle, attr):
+        elt = self.handle_to_node[handle]
+        return elt.attributes.get(attr, None)
+
+    def js_innerHTML(self, handle, s):
+        try:
+            self.run_rendering_pipeline()
+            doc = parse(lex("<!doctype><html><body>" +
+                            s + "</body></html>"))
+            new_nodes = doc.children[0].children
+            elt = self.handle_to_node[handle]
+            elt.children = new_nodes
+            for child in elt.children:
+                child.parent = elt
+            if self.document:
+                self.set_needs_reflow(
+                    layout_for_node(self.document, elt))
+            else:
+                self.set_needs_layout_tree_rebuild()
+        except:
+            import traceback
+            traceback.print_exc()
+            raise
+
+    def js_requestAnimationFrame(self):
+        self.needs_raf_callbacks = True
+        self.set_needs_animation_frame()
+
+    def js_now(self):
+        return int(time.time() * 1000)
+
+    def dispatch_event(self, type, elt):
+       handle = self.node_to_handle.get(elt, -1)
+
+       do_default = self.js.evaljs("__runHandlers({}, \"{}\")".format(handle, type))
+       return not do_default
+
+    def make_handle(self, elt):
+        if elt not in self.node_to_handle:
+            handle = len(self.node_to_handle)
+            self.node_to_handle[elt] = handle
+            self.handle_to_node[handle] = elt
         else:
-            self.focus = "content"
-            self.tabs[self.active_tab].click(e.x, e.y - CHROME_PX)
-        self.draw()
+            handle = self.node_to_handle[elt]
+        return handle
 
-    def handle_key(self, char):
-        if not (0x20 <= ord(char) < 0x7f): return
-        if self.focus == "address bar":
-            self.address_bar += char
-            self.draw()
-        elif self.focus == "content":
-            self.tabs[self.active_tab].keypress(char)
-            self.draw()
+    def set_needs_reflow(self, layout_object):
+        self.reflow_roots.append(layout_object)
+        self.set_needs_animation_frame()
 
-    def handle_enter(self):
-        if self.focus == "address bar":
-            self.tabs[self.active_tab].load(self.address_bar)
-            self.focus = None
-            self.draw()
+    def set_needs_layout_tree_rebuild(self):
+        self.needs_layout_tree_rebuild = True
+        self.set_needs_animation_frame()
 
-    def load(self, url):
-        new_tab = Tab()
-        new_tab.load(url)
-        self.active_tab = len(self.tabs)
-        self.tabs.append(new_tab)
-        self.draw()
+    def set_needs_animation_frame(self):
+        self.compositor_lock.acquire(blocking=True)
+        if not self.display_scheduled:
+            self.needs_animation_frame = True
+        self.compositor_lock.release()
+
+    def quit(self):
+        self.compositor_lock.acquire(blocking=True)
+        self.needs_quit = True
+        self.compositor_lock.release();        
+
+    def run_animation_frame(self):
+        self.needs_animation_frame = False
+
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.reset()
+
+        if (self.needs_raf_callbacks):
+            self.needs_raf_callbacks = False
+            if args.compute_main_thread_timings:
+                self.main_thread_timer.start("runRAFHandlers")
+            self.js.evaljs("__runRAFHandlers()")
+
+        self.run_rendering_pipeline()
+        # This will cause a draw to the screen, even if there are pending
+        # requestAnimationFrame callbacks for the *next* frame (which may have
+        # been registered during a call to __runRAFHandlers). By default,
+        # tkinter doesn't run these until there are no more event queue
+        # tasks.
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.start("IdleTasks")
+        self.canvas.update_idletasks()
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.stop()
+            self.main_thread_timer.print_accumulated()
+
+        self.frame_count = self.frame_count + 1
+        if args.stop_after > 0  and self.frame_count > args.stop_after:
+            self.quit()
+            sys.exit()
+
+    def run_rendering_pipeline(self):
+        if self.needs_layout_tree_rebuild:
+            self.document = DocumentLayout(self.nodes)
+            self.reflow_roots = [self.document]
+        self.needs_layout_tree_rebuild = False
+
+        for reflow_root in self.reflow_roots:
+            self.reflow(reflow_root)
+        self.reflow_roots = []
+        self.paint()
+        self.max_y = self.document.h - HEIGHT
+        # drawLayoutTree(self.document)
+
+    def paint(self):
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.start("Paint")
+        self.display_list = []
+        self.document.paint(self.display_list)
+
+    def reflow(self, obj):
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.start("Style")
+        style(obj.node, None, self.rules)
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.start("Layout (phase 1A)")
+        obj.size()
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.start("Layout (phase 1B)")
+        while obj.parent:
+            obj.parent.compute_height()
+            obj = obj.parent
+        if args.compute_main_thread_timings:
+            self.main_thread_timer.start("Layout (phase 2)")
+        self.document.position()
 
     def draw(self):
-        rasterizer = Rasterizer(self.skia_surface)
-        rasterizer.clear(skia.ColorWHITE)
-
-        self.tabs[self.active_tab].draw(rasterizer)
-    
-        # Draw the tabs UI:
-        tabfont = skia.Font(skia.Typeface('Arial'), 20)
-        for i, tab in enumerate(self.tabs):
-            name = "Tab {}".format(i)
-            x1, x2 = 40 + 80 * i, 120 + 80 * i
-            rasterizer.draw_polyline(x1, 0, x1, 40)
-            rasterizer.draw_polyline(x2, 0, x2, 40)
-            rasterizer.draw_text(x1 + 10, 10, name, tabfont)
-            if i == self.active_tab:
-                rasterizer.draw_polyline(0, 40, x1, 40)
-                rasterizer.draw_polyline(x2, 40, WIDTH, 40)
-
-        # Draw the plus button to add a tab:
-        buttonfont = skia.Font(skia.Typeface('Arial'), 30)
-        rasterizer.draw_rect(skia.Rect.MakeLTRB(10, 10, 30, 30))
-        rasterizer.draw_text(11, 0, "+", buttonfont)
-
-        # Draw the URL address bar:
-        rasterizer.draw_rect(
-            skia.Rect.MakeLTRB(40, 50, WIDTH - 10, 90))
+        if args.compute_compositor_thread_timings:
+            self.compositor_thread_timer.reset()
+            self.compositor_thread_timer.start("Draw")
+        self.canvas.delete("all")
+        for cmd in self.draw_display_list:
+            if cmd.y1 > self.scroll + HEIGHT - 60: continue
+            if cmd.y2 < self.scroll: continue
+            cmd.draw(self.scroll - 60, self.canvas)
+        if args.compute_compositor_thread_timings:
+            self.main_thread_timer.start("Draw Chrome")
+        self.canvas.create_rectangle(0, 0, 800, 60, width=0, fill='light gray')
+        self.canvas.create_rectangle(50, 10, 790, 50)
+        font = tkinter.font.Font(family="Courier", size=30)
+        self.canvas.create_text(55, 15, anchor='nw', text=self.address_bar, font=font)
+        self.canvas.create_rectangle(10, 10, 35, 50)
+        self.canvas.create_polygon(15, 30, 30, 15, 30, 45, fill='black')
         if self.focus == "address bar":
-            rasterizer.draw_text(55, 55, self.address_bar, buttonfont)
-            w = buttonfont.measureText(self.address_bar)
-            rasterizer.draw_polyline(55 + w, 55, 55 + w, 85)
-        else:
-            url = self.tabs[self.active_tab].url
-            rasterizer.draw_text(55, 55, url, buttonfont)
+            w = font.measure(self.address_bar)
+            self.canvas.create_line(55 + w, 15, 55 + w, 45)
+        elif isinstance(self.focus, InputLayout):
+            text = self.focus.node.attributes.get("value", "")
+            x = self.focus.x + self.focus.font.measure(text)
+            y = self.focus.y - self.scroll + 60
+            self.canvas.create_line(x, y, x, y + self.focus.h)
+        if args.compute_compositor_thread_timings:
+            self.compositor_thread_timer.stop()
+            self.compositor_thread_timer.print_accumulated()
 
-        # Draw the back button:
-        rasterizer.draw_rect(skia.Rect.MakeLTRB(10, 50, 35, 90))
-        rasterizer.draw_polyline(
-            15, 70, 30, 55, 30, 85, True)
-
-        # Raster the results and copy to the SDL surface:
-        skia_image = self.skia_surface.makeImageSnapshot()
-        skia_bytes = skia_image.tobytes()
-        rect = SDL_Rect(0, 0, WIDTH, HEIGHT)
-        skia_surface = Browser.to_sdl_surface(skia_bytes)
-        SDL_BlitSurface(
-            skia_surface, rect, self.window_surface, rect)
-        SDL_UpdateWindowSurface(self.sdl_window)
+    # Runs on the compositor thread
+    def scrolldown(self, e):
+        self.compositor_lock.acquire(blocking=True)
+        self.scroll = self.scroll + SCROLL_STEP
+        self.scroll = min(self.scroll, self.max_y)
+        self.scroll = max(0, self.scroll)
+        self.compositor_lock.release()
+        self.set_needs_animation_frame()
 
 if __name__ == "__main__":
     import sys
 
-    SDL_Init(SDL_INIT_VIDEO)
-    sdl_window = SDL_CreateWindow(b"Browser",
-        SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED,
-        WIDTH, HEIGHT, SDL_WINDOW_SHOWN)
+    parser = argparse.ArgumentParser(description="Chapter 13 source code")
+    parser.add_argument("--url", default=2, type=str, required=True,
+        help="URL to load")
+    parser.add_argument("--stop_after", default=0, type=int,
+        help="If set, exits the browser after this many generates frames")
+    parser.add_argument("--compute_main_thread_timings", type=bool,
+        help="Compute main thread timings")
+    parser.add_argument("--compute_compositor_thread_timings", type=bool,
+        help="Compute compositor thread timings")
+    args = parser.parse_args()
 
-    browser = Browser(sdl_window)
-    browser.load(sys.argv[1])
-
-    running = True
-    event = SDL_Event()
-    while running:
-        while SDL_PollEvent(ctypes.byref(event)) != 0:
-            if event.type == SDL_MOUSEMOTION or event.type == SDL_WINDOWEVENT:
-                continue
-            if event.type == SDL_MOUSEBUTTONUP:
-                browser.handle_click(event.button)
-            if event.type == SDL_KEYDOWN:
-                if event.key.keysym.sym == SDLK_RETURN:
-                    browser.handle_enter()
-                if event.key.keysym.sym == SDLK_DOWN:
-                    browser.handle_down()
-            if event.type == SDL_TEXTINPUT:
-                browser.handle_key(event.text.text.decode('utf8'))
-            if event.type == SDL_QUIT:
-                running = False
-                break
-
-    SDL_DestroyWindow(sdl_window)
-    SDL_Quit()
+    browser = Browser()
+    browser.start()
+    browser.schedule_load(args.url)
+    tkinter.mainloop()
