+++ src/lab13.py
@@ -1,28 +1,21 @@
 """
 This file compiles the code in Web Browser Engineering,
-up to and including Chapter 11 (Keeping Data Private),
+up to and including Chapter 11 (Saving Partial Layouts),
 without exercises.
 """
 
-import argparse
-import dukpy
-import functools
 import socket
 import ssl
-import time
-import threading
 import tkinter
 import tkinter.font
+import dukpy
+import time
 from lab10 import request
 
 class Timer:
     def __init__(self):
         self.phase = None
         self.time = None
-        self.accumulated = 0
-
-    def reset(self):
-        self.accumulated = 0
 
     def start(self, name):
         if self.phase: self.stop()
@@ -33,13 +26,6 @@ class Timer:
         dt = time.time() - self.time
         print("[{:>10.6f}] {}".format(dt, self.phase))
         self.phase = None
-        self.accumulated += dt
-
-    def print_accumulated(self):
-        print("[{:>10.6f}] {}\n".format(self.accumulated, "Total"))
-
-def url_origin(url):
-    return "/".join(url.split("/")[:3])
 
 class Text:
     def __init__(self, text):
@@ -93,7 +79,8 @@ class ElementNode:
     def __init__(self, tag, attributes):
         self.tag = tag
         self.attributes = attributes
-        self.children = [] 
+        self.children = []
+
         self.style = {}
         for pair in self.attributes.get("style", "").split(";"):
             if ":" not in pair: continue
@@ -254,7 +241,7 @@ class CSSParser:
     def parse(self):
         rules, _ = self.file(0)
         return rules
-
+    
 class TagSelector:
     def __init__(self, tag):
         self.tag = tag
@@ -335,10 +322,6 @@ class LineLayout:
     def compute_height(self):
         if not self.children:
             self.h = 0
-            self.max_ascent = 0
-            self.max_descent = 0
-            self.metrics = None
-            self.cxs = []
             return
         self.metrics = [child.font.metrics() for child in self.children]
         self.max_ascent = max([metric["ascent"] for metric in self.metrics])
@@ -353,42 +336,28 @@ class LineLayout:
 
     def position(self):
         baseline = self.y + 1.25 * self.max_ascent
-        if self.children:
-            for cx, child, metrics in \
-              zip(self.cxs, self.children, self.metrics):
-                child.x = self.x + cx
-                child.y = baseline - metrics["ascent"]
+        for cx, child, metrics in \
+          zip(self.cxs, self.children, self.metrics):
+            child.x = self.x + cx
+            child.y = baseline - metrics["ascent"]
 
-    def paint(self, to):
+    def draw(self, to):
         for child in self.children:
-            child.paint(to)
-
-FONT_CACHE = {}
-
-def GetFont(size, weight, style):
-    key = (size, weight, style)
-    value = FONT_CACHE.get(key)
-    if value: return value
-    value = tkinter.font.Font(size=size, weight=weight, slant=style)
-    FONT_CACHE[key] = value
-    return value
+            child.draw(to)
 
 class TextLayout:
     def __init__(self, node, word):
         self.node = node
         self.children = []
         self.word = word
-        self.display_item = None
 
     def size(self):
-        self.display_item = None
-
         weight = self.node.style["font-weight"]
         style = self.node.style["font-style"]
         if style == "normal": style = "roman"
         size = int(px(self.node.style["font-size"]) * .75)
-        self.font = GetFont(size, weight, style) 
-
+        self.font = tkinter.font.Font(size=size, weight=weight, slant=style)
+        
         self.w = self.font.measure(self.word)
         self.compute_height()
 
@@ -398,11 +367,9 @@ class TextLayout:
     def position(self):
         pass
 
-    def paint(self, to):
-        if not self.display_item:
-            color = self.node.style["color"]
-            self.display_item = DrawText(self.x, self.y, self.word, self.font, color)
-        to.append(self.display_item)
+    def draw(self, to):
+        color = self.node.style["color"]
+        to.append(DrawText(self.x, self.y, self.word, self.font, color))
 
 class InputLayout:
     def __init__(self, node):
@@ -424,7 +391,7 @@ class InputLayout:
     def position(self):
         pass
 
-    def paint(self, to):
+    def draw(self, to):
         x1, x2 = self.x, self.x + self.w
         y1, y2 = self.y, self.y + self.h
         bgcolor = "light gray" if self.node.tag == "input" else "yellow"
@@ -503,9 +470,9 @@ class InlineLayout:
             child.position()
             cy += child.h
 
-    def paint(self, to):
+    def draw(self, to):
         for child in self.children:
-            child.paint(to)
+            child.draw(to)
 
 def px(s):
     if s.endswith("px"):
@@ -563,7 +530,7 @@ class BlockLayout:
         self.compute_height()
 
     def compute_height(self):
-        self.h = 0
+        self.h = self.pt + self.pb
         for child in self.children:
             self.h += child.mt + child.h + child.mb
 
@@ -571,19 +538,19 @@ class BlockLayout:
         self.y += self.mt
         self.x += self.ml
 
-        y = self.y
+        y = self.y + self.pt
         for child in self.children:
             child.x = self.x + self.pl + self.bl
             child.y = y
             child.position()
             y += child.mt + child.h + child.mb
 
-    def paint(self, to):
+    def draw(self, to):
         if self.node.tag == "pre":
             x2, y2 = self.x + self.w, self.y + self.h
             to.append(DrawRect(self.x, self.y, x2, y2, "gray"))
         for child in self.children:
-            child.paint(to)
+            child.draw(to)
 
 class DocumentLayout:
     def __init__(self, node):
@@ -618,8 +585,8 @@ class DocumentLayout:
         child.y = self.y = 0
         child.position()
 
-    def paint(self, to):
-        self.children[0].paint(to)
+    def draw(self, to):
+        self.children[0].draw(to)
 
 class DrawText:
     def __init__(self, x1, y1, text, font, color):
@@ -718,108 +685,10 @@ def is_link(node):
     return isinstance(node, ElementNode) \
         and node.tag == "a" and "href" in node.attributes
 
-def drawHTMLTree(node, indent=0):
-    print(" "*indent, type(node).__name__, " ", node, sep="")
-    for child in node.children:
-        drawHTMLTree(child, indent + 2)
-
-def drawLayoutTree(node, indent=0):
+def drawTree(node, indent=0):
     print(" "*indent, type(node).__name__, " ", node.node, sep="")
     for child in node.children:
-        drawLayoutTree(child, indent + 2)
-
-REFRESH_RATE_MS = 16 # 16ms
-
-class Task:
-    def __init__(self, task_code, arg1=None, arg2=None):
-        self.task_code = task_code
-        self.arg1 = arg1
-        self.arg2 = arg2
-        self.__name__ = "task"
-
-    def __call__(self):
-        if self.arg2:
-            self.task_code(self.arg1, self.arg2)
-        elif self.arg1:
-            self.task_code(self.arg1)
-        else:
-            self.task_code()
-        # Prevent it accidentally running twice.
-        self.task_code = None
-        self.arg1 = None
-        self.arg2 = None
-
-class TaskQueue:
-    def __init__(self, lock):
-        self.tasks = []
-        self.lock = lock
-
-    def add_task(self, task_code):
-        self.lock.acquire(blocking=True)
-        self.tasks.append(task_code)
-        self.lock.release()
-
-    def has_tasks(self):
-        self.lock.acquire(blocking=True)
-        retval = len(self.tasks) > 0
-        self.lock.release()
-        return retval
-
-    def get_next_task(self):
-        self.lock.acquire(blocking=True)
-        retval = self.tasks.pop(0)
-        self.lock.release()
-        return retval
-
-class MainThreadRunner:
-    def __init__(self, browser):
-        self.lock = threading.Lock()
-        self.browser = browser
-        self.needs_animation_frame = False
-        self.main_thread = threading.Thread(target=self.run, args=())
-        self.script_tasks = TaskQueue(self.lock)
-        self.browser_tasks = TaskQueue(self.lock)
-
-    def schedule_animation_frame(self):
-        self.lock.acquire(blocking=True)
-        self.needs_animation_frame = True
-        self.lock.release()
-
-    def schedule_script_task(self, script):
-        self.script_tasks.add_task(script)
-
-    def schedule_browser_task(self, callback):
-        self.browser_tasks.add_task(callback)
-
-    def schedule_event_handler():
-        pass
-
-    def start(self):
-        self.main_thread.start()
-
-    def run(self):
-        while True:
-            self.lock.acquire(blocking=True)
-            needs_animation_frame = self.needs_animation_frame
-            self.lock.release()
-            if needs_animation_frame:
-                browser.run_animation_frame()
-                self.browser.commit()
-
-            browser_method = None
-            if self.browser_tasks.has_tasks():
-                browser_method = self.browser_tasks.get_next_task()
-            if browser_method:
-                browser_method()
-
-            script = None
-            if self.script_tasks.has_tasks():
-                script = self.script_tasks.get_next_task()
-
-            if script:
-                script()
-
-            time.sleep(0.001) # 1ms
+        drawTree(child, indent + 2)
 
 class Browser:
     def __init__(self):
@@ -830,108 +699,49 @@ class Browser:
             height=HEIGHT
         )
         self.canvas.pack()
-        self.cookies = {}
 
         self.history = []
         self.focus = None
         self.address_bar = ""
         self.scroll = 0
+
+        self.timer = Timer()
+        self.window.bind("<Down>", self.scrolldown)
+        self.window.bind("<Button-1>", self.handle_click)
+        self.window.bind("<Key>", self.keypress)
+        self.window.bind("<Return>", self.pressenter)
         self.display_list = []
 
-        self.draw_display_list = []
-        self.needs_draw = False
-
-        self.document = None
-
-        self.main_thread_timer = Timer()
-        self.compositor_thread_timer = Timer()
-        self.window.bind("<Down>", self.scrolldown)
-        self.window.bind("<Button-1>", self.compositor_handle_click)
-        self.window.bind("<Key>", self.compositor_keypress)
-        self.window.bind("<Return>", self.press_enter)
-
-        self.reflow_roots = []
-        self.needs_layout_tree_rebuild = False
-        self.needs_animation_frame = False
-        self.display_scheduled = False
-        self.needs_raf_callbacks = False
-
-        self.frame_count = 0
-        self.compositor_lock = threading.Lock()
-
-        self.needs_quit = False
-
-
-    def commit(self):
-        self.compositor_lock.acquire(blocking=True)
-        self.needs_draw = True
-        self.draw_display_list = self.display_list.copy()
-        self.compositor_lock.release()
-
-    def start(self):
-        self.main_thread_runner = MainThreadRunner(self)
-        self.main_thread_runner.start()
-        self.canvas.after(1, self.maybe_draw)
-
-    def maybe_draw(self):
-        self.compositor_lock.acquire(blocking=True)
-        if self.needs_quit:
-            sys.exit()
-        if self.needs_animation_frame and not self.display_scheduled:
-            self.canvas.after(
-                REFRESH_RATE_MS,
-                self.main_thread_runner.schedule_animation_frame)
-            self.display_scheduled = True
-
-        if self.needs_draw:
-            self.draw()
-        self.needs_draw = False
-        self.compositor_lock.release()
-        self.canvas.after(1, self.maybe_draw)
-
-    # Runs on the compositor thread
-    def compositor_handle_click(self, e):
+    def handle_click(self, e):
         self.focus = None
-        if e.y < 60:
-            # Browser chrome clicks can be handled without the main thread...
+        if e.y < 60: # Browser chrome
             if 10 <= e.x < 35 and 10 <= e.y < 50:
                 self.go_back()
             elif 50 <= e.x < 790 and 10 <= e.y < 50:
                 self.focus = "address bar"
                 self.address_bar = ""
-                self.set_needs_animation_frame()
+                self.render()
         else:
-            # ...but not clicks within the web page contents area
-            self.main_thread_runner.schedule_browser_task(
-                Task(self.handle_click, e))
+            x, y = e.x, e.y + self.scroll - 60
+            obj = find_layout(x, y, self.document)
+            if not obj: return
+            elt = obj.node
+            if elt and self.dispatch_event("click", elt): return
+            while elt:
+                if isinstance(elt, TextNode):
+                    pass
+                elif is_link(elt):
+                    url = resolve_url(elt.attributes["href"], self.url)
+                    return self.load(url)
+                elif elt.tag == "input":
+                    elt.attributes["value"] = ""
+                    self.focus = obj
+                    return self.reflow(self.focus)
+                elif elt.tag == "button":
+                    self.submit_form(elt)
+                elt = elt.parent
 
-    # Runs on the main thread
-    def handle_click(self, e):
-        # Lock to check scroll, which is updated on the compositor thread.
-        self.compositor_lock.acquire(blocking=True)
-        x, y = e.x, e.y + self.scroll - 60
-        self.compositor_lock.release()
-        self.run_rendering_pipeline()
-        obj = find_layout(x, y, self.document)
-        if not obj: return
-        elt = obj.node
-        if elt and self.dispatch_event("click", elt): return
-        while elt:
-            if isinstance(elt, TextNode):
-                pass
-            elif is_link(elt):
-                url = resolve_url(elt.attributes["href"], self.url)
-                return self.schedule_load(url)
-            elif elt.tag == "input":
-                elt.attributes["value"] = ""
-                self.focus = obj
-                self.set_needs_reflow(self.focus)
-            elif elt.tag == "button":
-                self.submit_form(elt)
-            elt = elt.parent
-
-    # Runs on the compositor thread
-    def compositor_keypress(self, e):
+    def keypress(self, e):
         if len(e.char) == 0: return
         if not (0x20 <= ord(e.char) < 0x7f): return
 
@@ -939,16 +749,11 @@ class Browser:
             return
         elif self.focus == "address bar":
             self.address_bar += e.char
-            self.set_needs_animation_frame()
+            self.render()
         else:
-            self.main_thread_runner.schedule_browser_task(
-                Task(self.keypress, e))
-
-    # Runs on the main thread
-    def keypress(self, e):
-        self.focus.node.attributes["value"] += e.char
-        self.dispatch_event("change", self.focus.node)
-        self.set_needs_reflow(self.focus)
+            self.focus.node.attributes["value"] += e.char
+            self.dispatch_event("change", self.focus.node)
+            self.reflow(self.focus)
 
     def submit_form(self, elt):
         while elt and elt.tag != "form":
@@ -963,86 +768,48 @@ class Browser:
             body += "&" + name + "=" + value.replace(" ", "%20")
         body = body[1:]
         url = resolve_url(elt.attributes["action"], self.url)
-        self.schedule_load(url, body)
+        self.load(url, body)
 
-    # Runs on the compositor thread
-    def press_enter(self, e):
+    def pressenter(self, e):
         if self.focus == "address bar":
             self.focus = None
-            self.schedule_load(self.address_bar)
+            self.load(self.address_bar)
 
     def go_back(self):
         if len(self.history) > 1:
             self.history.pop()
             back = self.history.pop()
-            self.schedule_load(back)
+            self.load(back)
 
-    def cookie_string(self):
-        cookie_string = ""
-        for key, value in self.cookies.items():
-            cookie_string += "&" + key + "=" + value
-        return cookie_string[1:]
-
-    # Runs on the compositor thread
-    def schedule_load(self, url, body=None):
-        self.main_thread_runner.schedule_browser_task(
-            Task(self.load, url, body))
-
-    # Runs on the main thread
     def load(self, url, body=None):
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.start("Downloading")
+        self.timer.start("Downloading")
         self.address_bar = url
         self.url = url
         self.history.append(url)
-        req_headers = { "Cookie": self.cookie_string() }
-        headers, body = request(url, headers=req_headers, payload=body)
-        if "set-cookie" in headers:
-            kv, params = headers["set-cookie"].split(";", 1)
-            key, value = kv.split("=", 1)
-            origin = url_origin(self.history[-1])
-            self.cookies.setdefault(origin, {})[key] = value
-
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.start("Parsing HTML")
+        header, body = request(url, payload=body)
+        self.timer.start("Parsing HTML")
         self.nodes = parse(lex(body))
-#        drawHTMLTree(self.nodes)
         
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.start("Parsing CSS")
+        self.timer.start("Parsing CSS")
         with open("browser8.css") as f:
             self.rules = CSSParser(f.read()).parse()
 
         for link in find_links(self.nodes, []):
-            header, body = request(resolve_url(link, url), headers=req_headers)
+            header, body = request(resolve_url(link, url))
             self.rules.extend(CSSParser(body).parse())
 
         self.rules.sort(key=lambda x: x[0].priority())
         self.rules.reverse()
 
-        self.run_scripts()
-        self.set_needs_layout_tree_rebuild()
-
-    def load_scripts(self, scripts):
-        req_headers = { "Cookie": self.cookie_string() }
-        for script in find_scripts(self.nodes, []):
-            header, body = request(
-                resolve_url(script, self.history[-1]), headers=req_headers)
-            scripts.append([header, body])
-
-    def script_run_wrapper(self, script_text):
-        return Task(self.js.evaljs, script_text)
-
-    def run_scripts(self):
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.start("Running JS")
+        self.timer.start("Running JS")
         self.setup_js()
-
-        scripts=[]
-        self.load_scripts(scripts)
-        for [header, body] in scripts:
-            self.main_thread_runner.schedule_script_task(
-                self.script_run_wrapper(body))
+        for script in find_scripts(self.nodes, []):
+            header, body = request(resolve_url(script, self.history[-1]))
+            try:
+                print("Script returned: ", self.js.evaljs(body))
+            except dukpy.JSRuntimeError as e:
+                print("Script", script, "crashed", e)
+        self.layout(self.nodes)
 
     def setup_js(self):
         self.js = dukpy.JSInterpreter()
@@ -1052,15 +819,8 @@ class Browser:
         self.js.export_function("querySelectorAll", self.js_querySelectorAll)
         self.js.export_function("getAttribute", self.js_getAttribute)
         self.js.export_function("innerHTML", self.js_innerHTML)
-        self.js.export_function("cookie", self.cookie_string)
-        self.js.export_function(
-            "requestAnimationFrame",
-            self.js_requestAnimationFrame)
-        self.js.export_function("now", self.js_now)
-        with open("runtime13.js") as f:
-            self.main_thread_runner.schedule_script_task(
-                self.script_run_wrapper(f.read()))
-
+        with open("runtime9.js") as f:
+            self.js.evaljs(f.read())
 
     def js_querySelectorAll(self, sel):
         selector, _ = CSSParser(sel + "{").selector(0)
@@ -1073,34 +833,20 @@ class Browser:
 
     def js_innerHTML(self, handle, s):
         try:
-            self.run_rendering_pipeline()
-            doc = parse(lex("<!doctype><html><body>" +
-                            s + "</body></html>"))
+            doc = parse(lex("<!doctype><html><body>" + s + "</body></html>"))
             new_nodes = doc.children[0].children
             elt = self.handle_to_node[handle]
             elt.children = new_nodes
             for child in elt.children:
                 child.parent = elt
-            if self.document:
-                self.set_needs_reflow(
-                    layout_for_node(self.document, elt))
-            else:
-                self.set_needs_layout_tree_rebuild()
+            self.reflow(layout_for_node(self.document, elt))
         except:
             import traceback
             traceback.print_exc()
             raise
 
-    def js_requestAnimationFrame(self):
-        self.needs_raf_callbacks = True
-        self.set_needs_animation_frame()
-
-    def js_now(self):
-        return int(time.time() * 1000)
-
     def dispatch_event(self, type, elt):
        handle = self.node_to_handle.get(elt, -1)
-
        do_default = self.js.evaljs("__runHandlers({}, \"{}\")".format(handle, type))
        return not do_default
 
@@ -1113,101 +859,35 @@ class Browser:
             handle = self.node_to_handle[elt]
         return handle
 
-    def set_needs_reflow(self, layout_object):
-        self.reflow_roots.append(layout_object)
-        self.set_needs_animation_frame()
-
-    def set_needs_layout_tree_rebuild(self):
-        self.needs_layout_tree_rebuild = True
-        self.set_needs_animation_frame()
-
-    def set_needs_animation_frame(self):
-        self.compositor_lock.acquire(blocking=True)
-        if not self.display_scheduled:
-            self.needs_animation_frame = True
-        self.compositor_lock.release()
-
-    def quit(self):
-        self.compositor_lock.acquire(blocking=True)
-        self.needs_quit = True
-        self.compositor_lock.release();        
-
-    def run_animation_frame(self):
-        self.needs_animation_frame = False
-
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.reset()
-
-        if (self.needs_raf_callbacks):
-            self.needs_raf_callbacks = False
-            if args.compute_main_thread_timings:
-                self.main_thread_timer.start("runRAFHandlers")
-            self.js.evaljs("__runRAFHandlers()")
-
-        self.run_rendering_pipeline()
-        # This will cause a draw to the screen, even if there are pending
-        # requestAnimationFrame callbacks for the *next* frame (which may have
-        # been registered during a call to __runRAFHandlers). By default,
-        # tkinter doesn't run these until there are no more event queue
-        # tasks.
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.start("IdleTasks")
-        self.canvas.update_idletasks()
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.stop()
-            self.main_thread_timer.print_accumulated()
-
-        self.frame_count = self.frame_count + 1
-        if args.stop_after > 0  and self.frame_count > args.stop_after:
-            self.quit()
-            sys.exit()
-
-    def run_rendering_pipeline(self):
-        if self.needs_layout_tree_rebuild:
-            self.document = DocumentLayout(self.nodes)
-            self.reflow_roots = [self.document]
-        self.needs_layout_tree_rebuild = False
-
-        for reflow_root in self.reflow_roots:
-            self.reflow(reflow_root)
-        self.reflow_roots = []
-        self.paint()
-        self.max_y = self.document.h - HEIGHT
-        # drawLayoutTree(self.document)
-
-    def paint(self):
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.start("Paint")
-        self.display_list = []
-        self.document.paint(self.display_list)
+    def layout(self, tree):
+        self.document = DocumentLayout(tree)
+        self.reflow(self.document)
 
     def reflow(self, obj):
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.start("Style")
-        style(obj.node, None, self.rules)
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.start("Layout (phase 1A)")
+        self.timer.start("Style")
+        style(obj.node, obj.node.parent, self.rules)
+        self.timer.start("Layout (phase 1A)")
         obj.size()
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.start("Layout (phase 1B)")
+        self.timer.start("Layout (phase 1B)")
         while obj.parent:
             obj.parent.compute_height()
             obj = obj.parent
-        if args.compute_main_thread_timings:
-            self.main_thread_timer.start("Layout (phase 2)")
+        self.timer.start("Layout (phase 2)")
         self.document.position()
+        self.timer.start("Display list")
+        self.display_list = []
+        self.document.draw(self.display_list)
+        self.render()
+        self.max_y = self.document.h - HEIGHT
 
-    def draw(self):
-        if args.compute_compositor_thread_timings:
-            self.compositor_thread_timer.reset()
-            self.compositor_thread_timer.start("Draw")
+    def render(self):
+        self.timer.start("Rendering")
         self.canvas.delete("all")
-        for cmd in self.draw_display_list:
+        for cmd in self.display_list:
             if cmd.y1 > self.scroll + HEIGHT - 60: continue
             if cmd.y2 < self.scroll: continue
             cmd.draw(self.scroll - 60, self.canvas)
-        if args.compute_compositor_thread_timings:
-            self.main_thread_timer.start("Draw Chrome")
+        self.timer.start("Chrome")
         self.canvas.create_rectangle(0, 0, 800, 60, width=0, fill='light gray')
         self.canvas.create_rectangle(50, 10, 790, 50)
         font = tkinter.font.Font(family="Courier", size=30)
@@ -1222,34 +902,16 @@ class Browser:
             x = self.focus.x + self.focus.font.measure(text)
             y = self.focus.y - self.scroll + 60
             self.canvas.create_line(x, y, x, y + self.focus.h)
-        if args.compute_compositor_thread_timings:
-            self.compositor_thread_timer.stop()
-            self.compositor_thread_timer.print_accumulated()
+        self.timer.stop()
 
-    # Runs on the compositor thread
     def scrolldown(self, e):
-        self.compositor_lock.acquire(blocking=True)
         self.scroll = self.scroll + SCROLL_STEP
         self.scroll = min(self.scroll, self.max_y)
         self.scroll = max(0, self.scroll)
-        self.compositor_lock.release()
-        self.set_needs_animation_frame()
+        self.render()
 
 if __name__ == "__main__":
     import sys
-
-    parser = argparse.ArgumentParser(description="Chapter 13 source code")
-    parser.add_argument("--url", default=2, type=str, required=True,
-        help="URL to load")
-    parser.add_argument("--stop_after", default=0, type=int,
-        help="If set, exits the browser after this many generates frames")
-    parser.add_argument("--compute_main_thread_timings", type=bool,
-        help="Compute main thread timings")
-    parser.add_argument("--compute_compositor_thread_timings", type=bool,
-        help="Compute compositor thread timings")
-    args = parser.parse_args()
-
     browser = Browser()
-    browser.start()
-    browser.schedule_load(args.url)
+    browser.load(sys.argv[1])
     tkinter.mainloop()
